+++ tests/conftest.py
@@ -28,7 +28,7 @@ def pytest_collection_modifyitems(items: list[pytest.Function]) -> None:
 
 base_url = os.environ.get("TEST_API_BASE_URL", "http://127.0.0.1:4010")
 
-auth_token = "My Auth Token"
+api_key = "My API Key"
 
 
 @pytest.fixture(scope="session")
@@ -37,7 +37,7 @@ def client(request: FixtureRequest) -> Iterator[BespokeLabs]:
     if not isinstance(strict, bool):
         raise TypeError(f"Unexpected fixture parameter type {type(strict)}, expected {bool}")
 
-    with BespokeLabs(base_url=base_url, auth_token=auth_token, _strict_response_validation=strict) as client:
+    with BespokeLabs(base_url=base_url, api_key=api_key, _strict_response_validation=strict) as client:
         yield client
 
 
@@ -47,5 +47,5 @@ async def async_client(request: FixtureRequest) -> AsyncIterator[AsyncBespokeLab
     if not isinstance(strict, bool):
         raise TypeError(f"Unexpected fixture parameter type {type(strict)}, expected {bool}")
 
-    async with AsyncBespokeLabs(base_url=base_url, auth_token=auth_token, _strict_response_validation=strict) as client:
+    async with AsyncBespokeLabs(base_url=base_url, api_key=api_key, _strict_response_validation=strict) as client:
         yield client